Update server_test to use string formatting, not decimal
Due to go vet being ran over the code, we need to ensure that the
correct type of string formatting is being used for types. This changes
the format for the `atomic.Int32` type in the server tests to use string
formatting, which is the correct type to use as `atomic.Int32`
implements the `fmt.Stringer` interface.