Update response.Problem to match RFC7808 more
This updates the response.Problem type to match RFC7808 more by adding
the instance field.

response.Problem now implements error, so a response.Problem can be
returned as an error by a type.

Also added is a ProblemError interface, this is to allow errors to be
asserted to be problems in order to fetch more error details.